Move AbstractLexicoder to API
